analysis: refactor driver tracking to separate helper class
